Add 'store' module and implement product display
This commit introduces the 'store' module. This includes the creation of new views, urls, templates and a 'Product' model in the module. Additionally, changes were made to the 'greatkart' project to integrate this new module. An extra context processor was added to ensure all categories are available across all templates. This update facilitates the display of available products according to categories.